requirements.txt
youtube-transcript-api
python-dotenv
streamlit>=1.28.0

# LangChain modular packages (no monorepo needed)
langchain-community>=0.2.12
langchain-openai>=0.2.0
langchain-core>=0.2.12
langchain-text-splitters>=0.2.2

# Vector store backend used by langchain_community.vectorstores.FAISS
faiss-cpu>=1.8.0

# Optional but useful for tokenization
tiktoken>=0.7.0